# API-CONTRACT.yaml
openapi: 3.1.0
info:
  title: Verdict API
  version: "1.0.0"
  description: >
    Canonical API contract for the Verdict hackathon judging platform.
    Every endpoint here must be implemented exactly as named — no additional
    endpoints invented by either the frontend or backend build without updating
    this file first. Enum values are defined once in ENUMS.md and referenced
    here by name. All list responses are wrapped in {"data": [...], "meta": {...}};
    all single-resource responses are wrapped in {"data": {...}}; all errors are
    {"error": {"code": "...", "message": "...", "details": {...}?}}.
servers:
  - url: /api/v1
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: verdict_session
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, enum: [unauthenticated, forbidden, not_found, validation_error, deadline_passed, event_not_open, conflict, rate_limited, internal_error] }
            message: { type: string }
            details: { type: object }
          required: [code, message]
      required: [error]
...
    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        event_id: { type: string, format: uuid }
        team_id: { type: string, format: uuid }
        track_id: { type: string, format: uuid, nullable: true }
        title: { type: string }
        summary: { type: string }
        description_md: { type: string, nullable: true }
        repo_url: { type: string, nullable: true }
        demo_url: { type: string, nullable: true }
        video_url: { type: string, nullable: true }
        cover_image_url: { type: string, nullable: true }
        status: { type: string, enum: [draft, submitted, disqualified] }
        submitted_at: { type: string, format: date-time, nullable: true }
        is_duplicate_of: { type: string, format: uuid, nullable: true }
...
  parameters:
    eventSlug:
      name: event_slug
      in: path
      required: true
      schema: { type: string }
    projectId:
      name: project_id
      in: path
      required: true
      schema: { type: string, format: uuid }
security:
  - sessionCookie: []
paths:
...
  /events/{event_slug}/teams:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    post:
      summary: Create a team (participant), caller becomes first member
      responses: { '201': { description: created }, '403': { description: submissions_closed or not authed } }
  /events/{event_slug}/teams/mine:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    get: { summary: Caller's team for this event, responses: { '200': { description: ok }, '404': { description: no team yet } } }
  /events/{event_slug}/teams/join:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    post:
      summary: Join a team via invite code
      requestBody: { content: { application/json: { schema: { type: object, properties: { code: { type: string } }, required: [code] } } } }
      responses: { '200': { description: joined }, '404': { description: invalid code }, '409': { description: expired or exhausted } }
  /teams/{team_id}/invites:
    parameters: [ { name: team_id, in: path, required: true, schema: { type: string, format: uuid } } ]
    post: { summary: Create an invite code (team member), responses: { '201': { description: created } } }
  /teams/{team_id}:
    parameters: [ { name: team_id, in: path, required: true, schema: { type: string, format: uuid } } ]
    get: { summary: Team detail, responses: { '200': { description: ok } } }

  # ---------------------------------------------------------------- PROJECTS (gallery / submit)
  /events/{event_slug}/projects:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    get:
      summary: >
        Public gallery. This is the .dogfood.toml `routes.gallery` target and must return JSON
        (not require JS execution) containing fixture project titles verbatim in response bodies.
      security: []
      parameters:
        - { name: q, in: query, schema: { type: string }, description: search title/summary }
        - { name: track_id, in: query, schema: { type: string, format: uuid } }
        - { name: sort, in: query, schema: { type: string, enum: [newest, title] } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Project' } }
                  meta: { type: object }
    post:
      summary: >
        Create + submit a project for the caller's team. This is the .dogfood.toml `routes.submit`
        target. Must return 4xx once events.submissions_close_at has passed, regardless of what the
        frontend UI shows.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                summary: { type: string }
                description_md: { type: string }
                repo_url: { type: string }
                demo_url: { type: string }
                video_url: { type: string }
                track_id: { type: string, format: uuid }
              required: [title, summary]
      responses:
        '201': { description: created }
        '403': { description: caller has no team / not a participant, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: submissions_closed — deadline enforced server-side, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /projects/{project_id}:
    parameters: [ { $ref: '#/components/parameters/projectId' } ]
    get: { summary: Project detail (public if submitted), security: [], responses: { '200': { description: ok }, '404': { description: not found } } }
    patch:
      summary: Edit a draft/submitted project (team member only, before deadline)
      responses:
        '200': { description: ok }
        '409': { description: deadline passed, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /projects/{project_id}/submit:
    parameters: [ { $ref: '#/components/parameters/projectId' } ]
    post:
      summary: Transition draft → submitted
      responses:
        '200': { description: ok }
        '409': { description: deadline passed, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ---------------------------------------------------------------- RUBRIC
  /events/{event_slug}/rubric:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    get: { summary: Active rubric, responses: { '200': { description: ok } } }
    put:
      summary: Replace active rubric configuration (organizer/admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                criteria:
                  type: array
                  items:
                    type: object
                    properties:
                      key: { type: string }
                      label: { type: string }
                      weight: { type: number }
                      scale_min: { type: integer }
                      scale_max: { type: integer }
                      option_labels: { type: object }
      responses: { '200': { description: ok }, '403': { description: forbidden } }

  # ---------------------------------------------------------------- ASSIGNMENTS
...